refactored the AutoCADData package, renaming it to 'data', and moving some of its classes to the grading package, where they are used
