+++ src/main/java/autocadDrawingChecker/data/elements/AbstractAutoCADText.java
@@ -1,4 +1,4 @@
-package autocadDrawingChecker.autocadData.elements;
+package autocadDrawingChecker.data.elements;
 
 /**
  *
