+++ src/main/java/autocadDrawingChecker/data/elements/AutoCADCircle.java
@@ -1,4 +1,4 @@
-package autocadDrawingChecker.autocadData.elements;
+package autocadDrawingChecker.data.elements;
 
 import java.util.Arrays;
