+++ src/main/java/autocadDrawingChecker/data/elements/AutoCADDiametricDimension.java
@@ -1,4 +1,4 @@
-package autocadDrawingChecker.autocadData.elements;
+package autocadDrawingChecker.data.elements;
 
 /**
  *
