+++ src/main/java/autocadDrawingChecker/data/extractors/CircleExtractor.java
@@ -1,7 +1,6 @@
-package autocadDrawingChecker.autocadData.extractors;
+package autocadDrawingChecker.data.extractors;
 
-import autocadDrawingChecker.autocadData.AutoCADAttribute;
-import autocadDrawingChecker.autocadData.elements.AutoCADCircle;
+import autocadDrawingChecker.data.elements.AutoCADCircle;
 
 /**
  *
