+++ src/main/java/autocadDrawingChecker/grading/implementations/CheckDims.java
@@ -1,7 +1,7 @@
 package autocadDrawingChecker.grading.implementations;
 
-import autocadDrawingChecker.autocadData.elements.AbstractAutoCADDimension;
-import autocadDrawingChecker.autocadData.elements.AutoCADElement;
+import autocadDrawingChecker.data.elements.AbstractAutoCADDimension;
+import autocadDrawingChecker.data.elements.AutoCADElement;
 import autocadDrawingChecker.grading.AbstractElementCriteria;
 import autocadDrawingChecker.grading.MathUtil;
