+++ src/main/java/autocadDrawingChecker/grading/implementations/LineLength.java
@@ -1,8 +1,8 @@
 package autocadDrawingChecker.grading.implementations;
 
-import autocadDrawingChecker.autocadData.AutoCADExport;
-import autocadDrawingChecker.autocadData.elements.AutoCADLine;
-import autocadDrawingChecker.autocadData.elements.AutoCADElement;
+import autocadDrawingChecker.data.elements.AutoCADExport;
+import autocadDrawingChecker.data.elements.AutoCADLine;
+import autocadDrawingChecker.data.elements.AutoCADElement;
 import autocadDrawingChecker.grading.AbstractElementCriteria;
 import autocadDrawingChecker.grading.MathUtil;
