+++ src/main/java/autocadDrawingChecker/start/Application.java
@@ -1,6 +1,6 @@
 package autocadDrawingChecker.start;
 
-import autocadDrawingChecker.autocadData.extractors.AbstractAutoCADElementExtractor;
+import autocadDrawingChecker.data.extractors.AbstractAutoCADElementExtractor;
 import autocadDrawingChecker.grading.AbstractGradingCriteria;
 import autocadDrawingChecker.grading.Grader;
 import autocadDrawingChecker.grading.GradingReport;
